Move tests too
