+++ javaagent-bootstrap/src/test/java/io/opentelemetry/javaagent/bootstrap/InternalJarURLHandlerTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.instrumentation.auto.api
+package io.opentelemetry.javaagent.bootstrap
 
 import io.opentelemetry.auto.util.test.AgentSpecification
 import io.opentelemetry.javaagent.bootstrap.InternalJarURLHandler
